+++ extensions/raycast-notification/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Raycast Notification Changelog
 
+## [New Features] - {PR_MERGE_DATE}
+
+- Add support for macOS Notification Center notifier (even the Reply-from-Notification feature)
+- Add more cross-extension usages
+- Add contributing steps
+- Bump all dependencies to the latest
+
 ## [Improvements] - 2024-08-20
 
 - Moved screenshots to metadata folder and used raycast colors for the command icon.